Adding a vertex type and changing to floats for colors
The vertex stores information about its position and its color,
within the vertex header the various color masking functions are also
defined, for working with packed uint8 colors. Changing mglColor
to take floats instead since that's the interface we'll present.

+++ include/microgl.h
@@ -41,7 +41,7 @@ void mgl_vertex3(float x, float y, float z);
 /*
 * Set the current color for vertices
 */
-void mgl_color(uint8_t r, uint8_t g, uint8_t b);
+void mgl_color(float r, float g, float b);
 /*
 * Set the current matrix mode
 */
@@ -94,3 +94,4 @@ void mgl_ortho(float left, float right, float bottom, float top,
 	float near, float far);
 
 #endif
+
